+++ CHANGELOG.md
@@ -4,7 +4,21 @@ Both client libraries are pre-1.0, and they have separate versioning.
 
 ## Unreleased
 
+- Added support for more parameters to `Sandbox.create()`:
+  - `blockNetwork`: Whether to block all network access from the sandbox.
+  - `cidrAllowlist`: List of CIDRs the sandbox is allowed to access.
+  - `gpu`: GPU reservation for the sandbox (e.g. "A100", "T4:2", "A100-80GB:4").
+  - `cloud`: Cloud provider to run the sandbox on.
+  - `regions`: Region(s) to run the sandbox on.
+  - `verbose`: Enable verbose logging.
+  - `proxy`: Connect a Modal Proxy to a Sandbox.
+  - `workdir`: Set the working directory.
 - Add top level for `Image.FromRawRegistry` in typescript and `ImageFromRawRegistry` in golang.
+
+## modal-js/v0.3.16, modal-go/v0.0.16
+
+- Added support for getting Sandboxes from an ID.
+- Added support for mounting `CloudBucketMount`s to Sandboxes.
 
 ## modal-js/v0.3.15, modal-go/v0.0.15